refactor: Add GET route for user verification
A new GET route has been added to the verify module in order to handle user verification. The route retrieves the user ID from the request URL parameters and checks if it exists in the database. If the user is found, an empty JSON response is returned. If the user is not found, a JSON response with a 404 status code is returned. This change improves the user verification process and enhances the overall functionality of the application.