refactor: Add redirect logic after successful login
This commit adds redirect logic to the login page component after a successful login. If a redirect query parameter is present in the URL, the user will be redirected to the specified page. Otherwise, the user will be redirected to the home page. This change improves the user experience by providing seamless navigation after login.