refactor: Update middleware to handle login page redirection
The middleware has been updated to handle the redirection logic for the login page. If the user is not logged in and is trying to access a non-API route, they will be redirected to the login page with the current path as a redirect query parameter. If the user is not logged in and is trying to access an API route, they will receive a JSON response with a 401 status code. Additionally, if the user is not logged in and is on the login page, the middleware will allow the request to proceed. This change improves the authentication flow and enhances the user experience.